upgrade dep-cruiser, fix violations
