+++ editor/package.json
@@ -325,7 +325,7 @@
     "concurrently": "3.5.1",
     "css-loader": "0.28.4",
     "csstype": "3.0.3",
-    "dependency-cruiser": "^9.17.1",
+    "dependency-cruiser": "13.1.5",
     "diff": "5.0.0",
     "enzyme": "3.3.0",
     "eslint": "7.32.0",
